@@ -1,5 +1,3 @@
-import 'dart:math';
-
 import 'package:flutter/widgets.dart';
 import 'package:html/dom.dart' as dom;
 
@@ -9,7 +7,6 @@ import 'core_widget_factory.dart';
 part 'data/build_bits.dart';
 part 'data/css.dart';
 part 'data/image.dart';
-part 'data/table.dart';
 part 'data/text_style.dart';
 
 /// A building element metadata.
@@ -17,10 +14,11 @@ abstract class BuildMetadata {
   /// The associatd element.
   final dom.Element element;
 
-  final TextStyleBuilder _tsb;
+  /// The associated [TextStyleBuilder].
+  final TextStyleBuilder tsb;
 
   /// Creates a node.
-  BuildMetadata(this.element, this._tsb);
+  BuildMetadata(this.element, this.tsb);
 
   /// The registered build ops.
   Iterable<BuildOp> get buildOps;
@@ -30,12 +28,12 @@ abstract class BuildMetadata {
 
   /// The inline styles.
   ///
-  /// These are usually collected from:
+  /// These are collected from:
   ///
   /// - [WidgetFactory.parse] or [BuildOp.onChild] by calling `meta[key] = value`
   /// - [BuildOp.defaultStyles] returning a map
   /// - Attribute `style` of [domElement]
-  Iterable<MapEntry<String, String>> get styles;
+  List<InlineStyle> get styles;
 
   /// Returns `true` if subtree will be built.
   ///
@@ -45,14 +43,14 @@ abstract class BuildMetadata {
   /// - Inline style `display: block`
   ///
   /// See [BuildOp.onWidgetsIsOptional].
-  bool get willBuildSubtree;
+  bool? get willBuildSubtree;
 
   /// Adds an inline style.
   operator []=(String key, String value);
 
   /// Gets an inline style value by key.
-  String operator [](String key) {
-    String value;
+  String? operator [](String key) {
+    String? value;
     for (final x in styles) {
       if (x.key == key) value = x.value;
     }
@@ -63,22 +61,17 @@ abstract class BuildMetadata {
   void register(BuildOp op);
 
   @override
-  String toString() =>
-      'BuildMetadata(${element == null ? "root" : element.outerHtml})';
-
-  /// Enqueues a text style builder callback.
-  ///
-  /// Returns the associated [TextStyleBuilder].
-  TextStyleBuilder tsb<T>([
-    TextStyleHtml Function(TextStyleHtml tsh, T input) builder,
-    T input,
-  ]) =>
-      _tsb..enqueue(builder, input);
+  String toString() => 'BuildMetadata(${element.outerHtml})';
 }
 
 /// A building operation to customize how a DOM element is rendered.
 @immutable
 class BuildOp {
+  /// The recommended maximum value for [priority].
+  ///
+  /// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/MAX_SAFE_INTEGER
+  static const kPriorityMax = 9007199254740991;
+
   /// The execution priority, op with lower priority will run first.
   ///
   /// Default: 10.
@@ -97,7 +90,7 @@ class BuildOp {
   ///
   /// Note: op must be registered early for this to work e.g.
   /// in [WidgetFactory.parse] or [onChild].
-  final Map<String, String> Function(dom.Element element) defaultStyles;
+  final Map<String, String> Function(dom.Element element)? defaultStyles;
 
   /// The callback that will be called whenver a child element is found.
   ///
@@ -113,16 +106,16 @@ class BuildOp {
   /// );
   ///
   /// ```
-  final void Function(BuildMetadata childMeta) onChild;
+  final void Function(BuildMetadata childMeta)? onChild;
 
   /// The callback that will be called when child elements have been processed.
-  final void Function(BuildMetadata meta, BuildTree tree) onTree;
+  final void Function(BuildMetadata meta, BuildTree tree)? onTree;
 
   /// The callback that will be called when child elements have been built.
   ///
   /// Note: only works if it's a block element.
-  final Iterable<Widget> Function(
-      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets) onWidgets;
+  final Iterable<Widget>? Function(
+      BuildMetadata meta, Iterable<WidgetPlaceholder> widgets)? onWidgets;
 
   /// Controls whether the element should be forced to be rendered as block.
   ///
