@@ -1,4 +1,5 @@
-import 'package:flutter/widgets.dart';
+import 'dart:collection';
+
 import 'package:html/dom.dart' as dom;
 
 import '../core_data.dart';
@@ -6,8 +7,6 @@ import '../core_data.dart' as core_data show BuildMetadata, BuildTree;
 import '../core_helpers.dart';
 import '../core_widget_factory.dart';
 import 'core_ops.dart';
-import 'core_parser.dart';
-import 'flattener.dart';
 
 final _regExpSpaceLeading = RegExp(r'^[^\S\u{00A0}]+', unicode: true);
 final _regExpSpaceTrailing = RegExp(r'[^\S\u{00A0}]+$', unicode: true);
@@ -16,71 +15,50 @@ final _regExpSpaces = RegExp(r'[^\S\u{00A0}]+', unicode: true);
 class BuildMetadata extends core_data.BuildMetadata {
   final Iterable<BuildOp> _parentOps;
 
-  List<BuildOp> _buildOps;
+  Set<BuildOp>? _buildOps;
   var _buildOpsIsLocked = false;
-  List<String> _styles;
+  List<InlineStyle>? _styles;
   var _stylesIsLocked = false;
-  bool _willBuildSubtree;
+  bool? _willBuildSubtree;
 
-  BuildMetadata(dom.Element element, TextStyleBuilder tsb, [this._parentOps])
+  BuildMetadata(dom.Element element, TextStyleBuilder tsb,
+      [this._parentOps = const []])
       : super(element, tsb);
 
   @override
-  Iterable<BuildOp> get buildOps => _buildOps;
+  Iterable<BuildOp> get buildOps => _buildOps ?? const [];
 
   @override
   Iterable<BuildOp> get parentOps => _parentOps;
 
   @override
-  Iterable<MapEntry<String, String>> get styles sync* {
+  List<InlineStyle> get styles {
     assert(_stylesIsLocked);
-
-    final iterator = _styles?.iterator;
-    while (iterator?.moveNext() == true) {
-      final key = iterator.current;
-      if (!iterator.moveNext()) return;
-      yield MapEntry(key, iterator.current);
-    }
+    return _styles ?? const [];
   }
 
   @override
-  bool get willBuildSubtree => _willBuildSubtree;
+  bool? get willBuildSubtree => _willBuildSubtree;
 
   @override
   operator []=(String key, String value) {
-    if (key == null || value == null) return;
-
     assert(!_stylesIsLocked, 'Metadata can no longer be changed.');
     _styles ??= [];
-    _styles..addAll([key, value]);
+    _styles!.add(InlineStyle(key, value));
   }
 
   @override
   void register(BuildOp op) {
-    if (op == null) return;
-
     assert(!_buildOpsIsLocked, 'Metadata can no longer be changed.');
-    _buildOps ??= [];
-    if (!buildOps.contains(op)) _buildOps.add(op);
-  }
-
-  void _sortBuildOps() {
-    assert(!_buildOpsIsLocked);
-
-    if (_buildOps != null) {
-      _buildOps.sort((a, b) => a.priority.compareTo(b.priority));
-    }
-
-    _willBuildSubtree = this[kCssDisplay] == kCssDisplayBlock ||
-        _buildOps?.where(_opRequiresBuildingSubtree)?.isNotEmpty == true;
-    _buildOpsIsLocked = true;
+    _buildOps ??= SplayTreeSet(_compareBuildOps);
+    _buildOps!.add(op);
   }
 }
 
 class BuildTree extends core_data.BuildTree {
-  final CustomStylesBuilder customStylesBuilder;
-  final CustomWidgetBuilder customWidgetBuilder;
-  final BuildMetadata parentMeta;
+  final CustomStylesBuilder? customStylesBuilder;
+  final CustomWidgetBuilder? customWidgetBuilder;
+  final core_data.BuildMetadata parentMeta;
   final Iterable<BuildOp> parentOps;
   final WidgetFactory wf;
 
@@ -89,15 +67,15 @@ class BuildTree extends core_data.BuildTree {
   BuildTree({
     this.customStylesBuilder,
     this.customWidgetBuilder,
-    BuildTree parent,
-    this.parentMeta,
-    this.parentOps,
-    TextStyleBuilder tsb,
-    this.wf,
+    core_data.BuildTree? parent,
+    required this.parentMeta,
+    this.parentOps = const [],
+    required TextStyleBuilder tsb,
+    required this.wf,
   }) : super(parent, tsb);
 
   @override
-  BuildBit add(BuildBit bit) {
+  T add<T extends BuildBit>(T bit) {
     assert(_built.isEmpty, "Built tree shouldn't be altered.");
     return super.add(bit);
   }
@@ -106,11 +84,8 @@ class BuildTree extends core_data.BuildTree {
     for (final domNode in domNodes) {
       _addBitsFromNode(domNode);
     }
-
-    if (parentMeta?.buildOps != null) {
-      for (final op in parentMeta.buildOps) {
-        op.onTree?.call(parentMeta, this);
-      }
+    for (final op in parentMeta.buildOps) {
+      op.onTree?.call(parentMeta, this);
     }
   }
 
@@ -118,16 +93,13 @@ class BuildTree extends core_data.BuildTree {
   Iterable<WidgetPlaceholder> build() {
     if (_built.isNotEmpty) return _built;
 
-    var widgets = _flatten();
-
-    if (parentMeta?.buildOps != null) {
-      for (final op in parentMeta.buildOps) {
-        widgets = op.onWidgets
-                ?.call(parentMeta, widgets)
-                ?.map(WidgetPlaceholder.lazy)
-                ?.toList(growable: false) ??
-            widgets;
-      }
+    var widgets = wf.flatten(parentMeta, this);
+    for (final op in parentMeta.buildOps) {
+      widgets = op.onWidgets
+              ?.call(parentMeta, widgets)
+              ?.map(WidgetPlaceholder.lazy)
+              .toList(growable: false) ??
+          widgets;
     }
 
     _built.addAll(widgets);
@@ -136,10 +108,10 @@ class BuildTree extends core_data.BuildTree {
 
   @override
   BuildTree sub({
-    core_data.BuildTree parent,
-    BuildMetadata parentMeta,
-    Iterable<BuildOp> parentOps,
-    TextStyleBuilder tsb,
+    core_data.BuildTree? parent,
+    BuildMetadata? parentMeta,
+    Iterable<BuildOp> parentOps = const [],
+    TextStyleBuilder? tsb,
   }) =>
       BuildTree(
         customStylesBuilder: customStylesBuilder,
@@ -152,7 +124,7 @@ class BuildTree extends core_data.BuildTree {
       );
 
   void _addBitsFromNode(dom.Node domNode) {
-    if (domNode.nodeType == dom.Node.TEXT_NODE) return _addText(domNode.text);
+    if (domNode.nodeType == dom.Node.TEXT_NODE) return _addText(domNode.text!);
     if (domNode.nodeType != dom.Node.ELEMENT_NODE) return;
 
     final element = domNode as dom.Element;
@@ -163,19 +135,19 @@ class BuildTree extends core_data.BuildTree {
       return;
     }
 
-    final meta = BuildMetadata(element, parentMeta.tsb().sub(), parentOps);
+    final meta = BuildMetadata(element, parentMeta.tsb.sub(), parentOps);
     _collectMetadata(meta);
 
     final subTree = sub(
       parentMeta: meta,
       parentOps: _prepareParentOps(parentOps, meta),
-      tsb: meta.tsb(),
+      tsb: meta.tsb,
     );
     add(subTree);
 
     subTree.addBitsFromNodes(element.nodes);
 
-    if (meta.willBuildSubtree) {
+    if (meta.willBuildSubtree == true) {
       for (final widget in subTree.build()) {
         add(WidgetBit.block(this, widget));
       }
@@ -206,34 +178,26 @@ class BuildTree extends core_data.BuildTree {
   void _collectMetadata(BuildMetadata meta) {
     wf.parse(meta);
 
-    if (meta.parentOps != null) {
-      for (final op in meta.parentOps) {
-        op.onChild?.call(meta);
-      }
+    for (final op in meta.parentOps) {
+      op.onChild?.call(meta);
     }
 
     // stylings, step 1: get default styles from tag-based build ops
-    if (meta.buildOps != null) {
-      for (final op in meta.buildOps) {
-        final map = op.defaultStyles?.call(meta.element);
-        if (map == null) continue;
+    for (final op in meta.buildOps) {
+      final map = op.defaultStyles?.call(meta.element);
+      if (map == null) continue;
 
-        meta._styles ??= [];
-        for (final pair in map.entries) {
-          if (pair.key == null || pair.value == null) continue;
-          meta._styles.insertAll(0, [pair.key, pair.value]);
-        }
+      meta._styles ??= [];
+      for (final pair in map.entries) {
+        meta._styles!.insert(0, InlineStyle(pair.key, pair.value));
       }
     }
 
     _customStylesBuilder(meta);
 
     // stylings, step 2: get styles from `style` attribute
-    final attrs = meta.element.attributes;
-    if (attrs.containsKey('style')) {
-      for (final pair in splitAttributeStyle(attrs['style'])) {
-        meta[pair.key] = pair.value;
-      }
+    for (final pair in meta.element.styles) {
+      meta[pair.key] = pair.value;
     }
 
     meta._stylesIsLocked = true;
@@ -243,7 +207,9 @@ class BuildTree extends core_data.BuildTree {
 
     wf.parseStyleDisplay(meta, meta[kCssDisplay]);
 
-    meta._sortBuildOps();
+    meta._willBuildSubtree = meta[kCssDisplay] == kCssDisplayBlock ||
+        meta._buildOps?.where(_opRequiresBuildingSubtree).isNotEmpty == true;
+    meta._buildOpsIsLocked = true;
   }
 
   void _customStylesBuilder(BuildMetadata meta) {
@@ -254,41 +220,19 @@ class BuildTree extends core_data.BuildTree {
       meta[pair.key] = pair.value;
     }
   }
+}
 
-  Iterable<WidgetPlaceholder> _flatten() {
-    final widgets = <WidgetPlaceholder>[];
+int _compareBuildOps(BuildOp a, BuildOp b) {
+  if (identical(a, b)) return 0;
 
-    for (final flattened in flatten(this)) {
-      if (flattened.widget != null) {
-        widgets.add(WidgetPlaceholder.lazy(flattened.widget));
-        continue;
-      }
-
-      if (flattened.widgetBuilder != null) {
-        widgets.add(WidgetPlaceholder<BuildTree>(this)
-            .wrapWith((context, _) => flattened.widgetBuilder(context)));
-        continue;
-      }
-
-      if (flattened.spanBuilder == null) continue;
-      widgets.add(WidgetPlaceholder<BuildTree>(this).wrapWith((context, _) {
-        final span = flattened.spanBuilder(context);
-        if (span == null || span is! InlineSpan) return widget0;
-
-        final tsh = tsb?.build(context);
-        final textAlign = tsh?.textAlign ?? TextAlign.start;
-
-        if (span is WidgetSpan &&
-            span.alignment == PlaceholderAlignment.baseline &&
-            textAlign == TextAlign.start) {
-          return span.child;
-        }
-
-        return wf.buildText(parentMeta, tsh, span);
-      }));
-    }
-
-    return widgets;
+  final cmp = a.priority.compareTo(b.priority);
+  if (cmp == 0) {
+    // if two ops have the same priority, they should not be considered equal
+    // fallback to compare hash codes for stable sorting
+    // while still provide pseudo random order across different runs
+    return a.hashCode.compareTo(b.hashCode);
+  } else {
+    return cmp;
   }
 }
 
@@ -297,9 +241,8 @@ bool _opRequiresBuildingSubtree(BuildOp op) =>
 
 Iterable<BuildOp> _prepareParentOps(Iterable<BuildOp> ops, BuildMetadata meta) {
   // try to reuse existing list if possible
-  final withOnChild =
-      meta.buildOps?.where((op) => op.onChild != null)?.toList();
-  return withOnChild?.isNotEmpty != true
+  final withOnChild = meta.buildOps.where((op) => op.onChild != null).toList();
+  return withOnChild.isEmpty
       ? ops
-      : List.unmodifiable([if (ops != null) ...ops, ...withOnChild]);
+      : List.unmodifiable([...ops, ...withOnChild]);
 }
