@@ -4,28 +4,30 @@ import 'package:flutter/widgets.dart';
 import '../core_data.dart';
 import 'margin_vertical.dart';
 
-List<Flattened> flatten(BuildTree tree) => _instance.flatten(tree);
-
 @immutable
 class Flattened {
-  final SpanBuilder spanBuilder;
-  final Widget widget;
-  final WidgetBuilder widgetBuilder;
+  final SpanBuilder? spanBuilder;
+  final Widget? widget;
+  final WidgetBuilder? widgetBuilder;
 
   Flattened._({this.spanBuilder, this.widget, this.widgetBuilder});
 }
 
-typedef SpanBuilder = InlineSpan Function(BuildContext);
+typedef SpanBuilder = InlineSpan? Function(BuildContext);
 
-final _instance = _Flattener();
+class Flattener {
+  final List<GestureRecognizer> _recognizers = [];
 
-class _Flattener {
-  List<Flattened> _flattened;
-  StringBuffer _buffer, _prevBuffer;
-  _Recognizer _recognizer, _prevRecognizer;
-  List<dynamic> _spans;
-  bool _swallowWhitespace;
-  TextStyleBuilder _tsb, _prevTsb;
+  late List<Flattened> _flattened;
+  late StringBuffer _buffer, _prevBuffer;
+  late _Recognizer _recognizer, _prevRecognizer;
+  List<dynamic>? _spans;
+  late bool _swallowWhitespace;
+  late TextStyleBuilder _tsb, _prevTsb;
+
+  void dispose() => _reset();
+
+  void reset() => _reset();
 
   List<Flattened> flatten(BuildTree tree) {
     _flattened = [];
@@ -37,6 +39,13 @@ class _Flattener {
     _completeLoop();
 
     return _flattened;
+  }
+
+  void _reset() {
+    for (final r in _recognizers) {
+      r.dispose();
+    }
+    _recognizers.clear();
   }
 
   void _resetLoop(TextStyleBuilder tsb) {
@@ -52,11 +61,9 @@ class _Flattener {
   }
 
   void _loop(final BuildBit bit) {
-    final tsb = _getBitTsb(bit);
-    if (_spans == null) _resetLoop(tsb);
-
-    final thisTsb = tsb ?? _prevTsb;
-    if (thisTsb?.hasSameStyleWith(_prevTsb) == false) _saveSpan();
+    final thisTsb = _getBitTsb(bit);
+    if (_spans == null) _resetLoop(thisTsb);
+    if (!thisTsb.hasSameStyleWith(_prevTsb)) _saveSpan();
 
     var built;
     if (bit is BuildBit<Null, dynamic>) {
@@ -65,10 +72,9 @@ class _Flattener {
       // ignore: omit_local_variable_types
       final WidgetBuilder widgetBuilder = (c) => bit.buildBit(c);
       built = widgetBuilder;
-    } else if (bit is BuildBit<GestureRecognizer, dynamic>) {
+    } else if (bit is BuildBit<GestureRecognizer?, dynamic>) {
       built = bit.buildBit(_prevRecognizer.value);
     } else if (bit is BuildBit<TextStyleHtml, InlineSpan>) {
-      // ignore: omit_local_variable_types
       final SpanBuilder spanBuilder = (c) => bit.buildBit(thisTsb.build(c));
       built = spanBuilder;
     }
@@ -77,7 +83,7 @@ class _Flattener {
       _prevRecognizer.value = built;
     } else if (built is InlineSpan || built is SpanBuilder) {
       _saveSpan();
-      _spans.add(built);
+      _spans!.add(built);
     } else if (built is String) {
       if (built != ' ' || !_loopShouldSwallowWhitespace(bit)) {
         _prevBuffer.write(built);
@@ -104,7 +110,7 @@ class _Flattener {
     // special handling for whitespaces
     if (_swallowWhitespace) return true;
 
-    final next = nextWithTsb(bit);
+    final next = nextNonWhitespace(bit);
     if (next == null) {
       // skip trailing whitespace
       return true;
@@ -121,9 +127,12 @@ class _Flattener {
       final scopedRecognizer = _prevRecognizer.value;
       final scopedTsb = _prevTsb;
       final scopedText = _prevBuffer.toString();
-      _spans.add((context) => TextSpan(
+
+      if (scopedRecognizer != null) _recognizers.add(scopedRecognizer);
+
+      _spans!.add((context) => TextSpan(
             recognizer: scopedRecognizer,
-            style: scopedTsb?.build(context)?.styleWithHeight,
+            style: scopedTsb.build(context).styleWithHeight,
             text: scopedText,
           ));
     }
@@ -134,14 +143,16 @@ class _Flattener {
   void _completeLoop() {
     _saveSpan();
 
-    if (_spans == null) return;
-
     final scopedSpans = _spans;
+    if (scopedSpans == null) return;
+
     _spans = null;
     if (scopedSpans.isEmpty && _buffer.isEmpty) return;
     final scopedRecognizer = _recognizer.value;
     final scopedTsb = _tsb;
     final scopedBuffer = _buffer.toString();
+
+    if (scopedRecognizer != null) _recognizers.add(scopedRecognizer);
 
     // trim the last new line if any
     final scopedText = scopedSpans.isEmpty
@@ -169,36 +180,36 @@ class _Flattener {
       return TextSpan(
         children: children,
         recognizer: scopedRecognizer,
-        style: scopedTsb?.build(context)?.styleWithHeight,
+        style: scopedTsb.build(context).styleWithHeight,
         text: scopedText,
       );
     }));
   }
 
-  static TextStyleBuilder _getBitTsb(BuildBit bit) {
-    if (bit.tsb != null) return bit.tsb;
+  TextStyleBuilder _getBitTsb(BuildBit bit) {
+    if (bit is! WhitespaceBit) return bit.tsb;
 
     // the below code will find the best style for this whitespace bit
     // easy case: whitespace at the beginning of a tag, use the previous style
     final parent = bit.parent;
-    if (parent == null || bit == parent.first) return null;
+    if (parent == null || bit == parent.first) return _prevTsb;
 
     // complicated: whitespace at the end of a tag, try to merge with the next
     // unless it has unrelated style (e.g. next bit is a sibling)
     if (bit == parent.last) {
-      final next = nextWithTsb(bit);
+      final next = nextNonWhitespace(bit);
       if (next != null) {
         var tree = parent;
         while (true) {
           final bitsParentLast = tree.parent?.last;
           if (bitsParentLast != bit) break;
-          tree = tree.parent;
+          tree = tree.parent!;
         }
 
         if (tree.parent == next.parent) {
           return next.tsb;
         } else {
-          return null;
+          return _prevTsb;
         }
       }
     }
@@ -207,9 +218,9 @@ class _Flattener {
     return parent.tsb;
   }
 
-  static BuildBit nextWithTsb(BuildBit bit) {
+  static BuildBit? nextNonWhitespace(BuildBit bit) {
     var next = bit.next;
-    while (next != null && next.tsb == null) {
+    while (next != null && next is WhitespaceBit) {
       next = next.next;
     }
 
@@ -222,5 +233,5 @@ class _Flattener {
 /// This class is needed because [GestureRecognizer] is rebuilt
 /// but we have to keep track of prev / current recognizer for spans
 class _Recognizer {
-  GestureRecognizer value;
+  GestureRecognizer? value;
 }
