@@ -11,6 +11,16 @@ const kTagTableCaption = 'caption';
 
 const kAttributeBorder = 'border';
 const kAttributeCellPadding = 'cellpadding';
+const kAttributeColspan = 'colspan';
+const kAttributeCellSpacing = 'cellspacing';
+const kAttributeRowspan = 'rowspan';
+const kAttributeValign = 'valign';
+
+const kCssBorderCollapse = 'border-collapse';
+const kCssBorderCollapseCollapse = 'collapse';
+const kCssBorderCollapseSeparate = 'separate';
+const kCssBorderSpacing = 'border-spacing';
+
 const kCssDisplayTable = 'table';
 const kCssDisplayTableRow = 'table-row';
 const kCssDisplayTableHeaderGroup = 'table-header-group';
@@ -20,33 +30,37 @@ const kCssDisplayTableCell = 'table-cell';
 const kCssDisplayTableCaption = 'table-caption';
 
 class TagTable {
+  final companion = HtmlTableCompanion();
   final BuildMetadata tableMeta;
   final WidgetFactory wf;
 
+  final _captions = <BuildTree>[];
   final _data = _TagTableData();
 
-  BuildOp _tableOp;
+  BuildOp? _tableOp;
 
   TagTable(this.wf, this.tableMeta);
 
   BuildOp get op {
-    _tableOp = BuildOp(
+    return _tableOp ??= BuildOp(
       onChild: onChild,
+      onTree: onTree,
       onWidgets: onWidgets,
-      priority: StyleSizing.kPriority,
+      priority: 0,
     );
-    return _tableOp;
   }
 
   void onChild(BuildMetadata childMeta) {
     if (childMeta.element.parent != tableMeta.element) return;
 
     final which = _getCssDisplayValue(childMeta);
+    _TagTableDataGroup? latestGroup;
     switch (which) {
       case kCssDisplayTableRow:
+        latestGroup ??= _data.body;
         final row = _TagTableDataRow();
-        _data.rows.add(row);
-        childMeta.register(_TagTableRow(wf, childMeta, row).op);
+        latestGroup.rows.add(row);
+        childMeta.register(_TagTableRow(this, childMeta, row).op);
         break;
       case kCssDisplayTableHeaderGroup:
       case kCssDisplayTableRowGroup:
@@ -54,78 +68,127 @@ class TagTable {
         final rows = which == kCssDisplayTableHeaderGroup
             ? _data.header.rows
             : which == kCssDisplayTableRowGroup
-                ? _data.rows
+                ? _data.body.rows
                 : _data.footer.rows;
-        childMeta.register(_TagTableGroup(wf, childMeta, rows).op);
+        childMeta.register(_TagTableRowGroup(this, childMeta, rows).op);
+        latestGroup = null;
         break;
       case kCssDisplayTableCaption:
-        childMeta.register(BuildOp(
-          onWidgets: (meta, widgets) {
-            final caption = wf
-                .buildColumnPlaceholder(meta, widgets)
-                ?.wrapWith((_, child) => _TableCaption(child));
-            if (caption == null) return [];
-
-            _data.captions.add(caption);
-            return [caption];
-          },
-          priority: StyleSizing.kPriority,
-        ));
+        childMeta.register(BuildOp(onTree: (_, tree) => _captions.add(tree)));
         break;
+    }
+  }
+
+  void onTree(BuildMetadata _, BuildTree tree) {
+    StyleBorder.skip(tableMeta);
+    StyleSizing.treatHeightAsMinHeight(tableMeta);
+
+    for (final caption in _captions) {
+      final built = wf
+          .buildColumnPlaceholder(tableMeta, caption.build())
+          ?.wrapWith((_, child) => _TableCaption(child));
+      if (built != null) {
+        WidgetBit.block(tree.parent!, built).insertBefore(tree);
+      }
+
+      caption.detach();
     }
   }
 
   Iterable<Widget> onWidgets(BuildMetadata _, Iterable<WidgetPlaceholder> __) {
-    final rows = <_TagTableDataRow>[
-      ..._data.header.rows,
-      ..._data.rows,
-      ..._data.footer.rows,
-    ];
+    final builders = <_HtmlTableCellBuilder>[];
+    // occupations data: { rowId: { columnId: occupied } }
+    final occupations = <int, Map<int, bool>>{};
+    _prepareHtmlTableCellBuilders(_data.header, occupations, builders);
+    for (final body in _data.bodies) {
+      _prepareHtmlTableCellBuilders(body, occupations, builders);
+    }
+    _prepareHtmlTableCellBuilders(_data.footer, occupations, builders);
+    if (builders.isEmpty) return [];
 
-    if (_data.captions.isEmpty && rows.isEmpty) return [];
+    final border = tryParseBorder(tableMeta);
+    final borderCollapse = tableMeta[kCssBorderCollapse];
+    final borderSpacing = tryParseCssLength(tableMeta[kCssBorderSpacing]);
 
     return [
       WidgetPlaceholder<BuildMetadata>(tableMeta).wrapWith((context, _) {
-        final metadata = TableMetadata(border: _parseBorder(context));
+        final tsh = tableMeta.tsb.build(context);
 
-        for (var i = 0; i < rows.length; i++) {
-          for (final cell in rows[i].cells) {
-            metadata.addCell(
-              i,
-              cell.child,
-              colspan: cell.colspan,
-              rowspan: cell.rowspan,
-            );
-          }
-        }
-
-        final tsh = tableMeta.tsb().build(context);
-        final table = wf.buildTable(tableMeta, tsh, metadata);
-        return wf.buildColumnWidget(tableMeta, tsh, [
-          ..._data.captions,
-          if (table != null) table,
-        ]);
+        return HtmlTable(
+          border: border.getValue(tsh),
+          borderCollapse: borderCollapse == kCssBorderCollapseCollapse,
+          borderSpacing: borderSpacing?.getValue(tsh) ?? 0.0,
+          companion: companion,
+          children: List.from(
+              builders
+                  .map((f) => f(context))
+                  .where((element) => element != null),
+              growable: false),
+        );
       }),
     ];
   }
 
-  BorderSide _parseBorder(BuildContext context) {
-    final value = tableMeta[kCssBorder];
-    if (value != null) {
-      final borderParsed = tryParseCssBorderSide(value);
-      if (borderParsed != null) {
-        return BorderSide(
-          color: borderParsed.color ?? const Color(0xFF000000),
-          width: borderParsed.width.getValue(tableMeta.tsb().build(context)),
-        );
+  void _prepareHtmlTableCellBuilders(
+      _TagTableDataGroup group,
+      Map<int, Map<int, bool>> occupations,
+      List<_HtmlTableCellBuilder> builders) {
+    var rowStart = occupations.keys.length - 1;
+    final rowSpanMax = group.rows.length;
+    for (final row in group.rows) {
+      rowStart++;
+      occupations[rowStart] ??= {};
+
+      for (final cell in row.cells) {
+        var columnStart = 0;
+        while (occupations[rowStart]!.containsKey(columnStart)) {
+          columnStart++;
+        }
+
+        final columnSpan = cell.colspan > 0 ? cell.colspan : 1;
+        final rowSpan = min(
+            rowSpanMax,
+            cell.rowspan > 0
+                ? cell.rowspan
+                : cell.rowspan == 0
+                    ? group.rows.length
+                    : 1);
+        for (var r = 0; r < rowSpan; r++) {
+          final row = rowStart + r;
+          occupations[row] ??= {};
+          for (var c = 0; c < columnSpan; c++) {
+            occupations[row]![columnStart + c] = true;
+          }
+        }
+
+        final cellMeta = cell.meta;
+        cellMeta.row = rowStart;
+
+        final cssBorderParsed = tryParseBorder(cellMeta);
+        final cssBorder = cssBorderParsed.inherit
+            ? tryParseBorder(tableMeta).copyFrom(cssBorderParsed)
+            : cssBorderParsed;
+
+        builders.add((context) {
+          Widget? child = cell.child;
+
+          final border = cssBorder.getValue(cellMeta.tsb.build(context));
+          if (border != null) {
+            child = wf.buildPadding(cellMeta, cell.child, border.dimensions);
+          }
+          if (child == null) return null;
+
+          return HtmlTableCell(
+            border: border,
+            columnSpan: columnSpan,
+            columnStart: columnStart,
+            rowSpan: rowSpan,
+            rowStart: cellMeta.row,
+            child: child,
+          );
+        });
       }
     }
-
-    final width =
-        tryParseDoubleFromMap(tableMeta.element.attributes, kAttributeBorder);
-    if (width != null && width > 0) return BorderSide(width: width);
-
-    return null;
   }
 
   static BuildOp cellPaddingOp(double px) => BuildOp(
@@ -134,8 +197,25 @@ class TagTable {
               ? meta[kCssPadding] = '${px}px'
               : null);
 
-  static String _getCssDisplayValue(BuildMetadata meta) {
-    String value;
+  static BuildOp borderOp(double border, double borderSpacing) => BuildOp(
+        defaultStyles: (_) => {
+          kCssBorder: '${border}px solid black',
+          kCssBorderCollapse: kCssBorderCollapseSeparate,
+          kCssBorderSpacing: '${borderSpacing}px',
+        },
+        onChild: border > 0
+            ? (meta) {
+                switch (meta.element.localName) {
+                  case kTagTableCell:
+                  case kTagTableHeaderCell:
+                    meta[kCssBorder] = kCssBorderInherit;
+                }
+              }
+            : null,
+      );
+
+  static String? _getCssDisplayValue(BuildMetadata meta) {
+    String? value;
     switch (meta.element.localName) {
       case kTagTableRow:
         value = kCssDisplayTableRow;
@@ -161,14 +241,9 @@ class TagTable {
       return value;
     }
 
-    final attrs = meta.element.attributes;
-    if (attrs.containsKey('style')) {
-      for (final pair in splitAttributeStyle(attrs['style'])
-          .toList(growable: false)
-          .reversed) {
-        if (pair.key == kCssDisplay) {
-          return pair.value;
-        }
+    for (final pair in meta.element.styles.reversed) {
+      if (pair.key == kCssDisplay) {
+        return pair.value;
       }
     }
 
@@ -176,28 +251,97 @@ class TagTable {
   }
 }
 
+extension _BuildMetadataExtension on BuildMetadata {
+  static final _rows = Expando<int>();
+
+  set row(int v) => _rows[this] = v;
+  int get row => _rows[this] ?? -1;
+}
+
+typedef _HtmlTableCellBuilder = HtmlTableCell? Function(BuildContext);
+
 class _TableCaption extends SingleChildRenderObjectWidget {
-  _TableCaption(Widget child, {Key key}) : super(child: child, key: key);
+  _TableCaption(Widget child, {Key? key}) : super(child: child, key: key);
 
   @override
   RenderObject createRenderObject(BuildContext context) => RenderProxyBox();
 }
 
-class _TableCell extends SingleChildRenderObjectWidget {
-  _TableCell(Widget child, {Key key}) : super(child: child, key: key);
+class _TagTableRow {
+  late final BuildOp op;
+  final TagTable parent;
+  final _TagTableDataRow row;
+  final BuildMetadata rowMeta;
 
-  @override
-  RenderObject createRenderObject(BuildContext context) => RenderProxyBox();
+  late final BuildOp _cellOp;
+  late final BuildOp _valignBaselineOp;
+
+  _TagTableRow(this.parent, this.rowMeta, this.row) {
+    op = BuildOp(onChild: onChild);
+    _cellOp = BuildOp(
+      onWidgets: (cellMeta, widgets) {
+        final column = parent.wf.buildColumnPlaceholder(cellMeta, widgets);
+        if (column == null) return [];
+
+        final attributes = cellMeta.element.attributes;
+        row.cells.add(_TagTableDataCell(
+          cellMeta,
+          child: column,
+          colspan: tryParseIntFromMap(attributes, kAttributeColspan) ?? 1,
+          rowspan: tryParseIntFromMap(attributes, kAttributeRowspan) ?? 1,
+        ));
+
+        return [column];
+      },
+      priority: BuildOp.kPriorityMax,
+    );
+    _valignBaselineOp = BuildOp(
+      onWidgets: (cellMeta, widgets) {
+        final v = cellMeta[kCssVerticalAlign];
+        if (v != kCssVerticalAlignBaseline) return widgets;
+
+        return listOrNull(parent.wf
+            .buildColumnPlaceholder(cellMeta, widgets)
+            ?.wrapWith((_, child) {
+          final row = cellMeta.row;
+
+          return HtmlTableValignBaseline(
+            companion: parent.companion,
+            row: row,
+            child: child,
+          );
+        }));
+      },
+      priority: StyleVerticalAlign.kPriority4500,
+    );
+  }
+
+  void onChild(BuildMetadata childMeta) {
+    if (childMeta.element.parent != rowMeta.element) return;
+    if (TagTable._getCssDisplayValue(childMeta) != kCssDisplayTableCell) {
+      return;
+    }
+
+    final attrs = childMeta.element.attributes;
+    if (attrs.containsKey(kAttributeValign)) {
+      childMeta[kCssVerticalAlign] = attrs[kAttributeValign]!;
+    }
+
+    childMeta.register(_cellOp);
+    StyleBorder.skip(childMeta);
+    StyleSizing.treatHeightAsMinHeight(childMeta);
+    childMeta.register(_valignBaselineOp);
+  }
 }
 
-class _TagTableGroup {
+class _TagTableRowGroup {
+  final TagTable parent;
   final List<_TagTableDataRow> rows;
   final BuildMetadata groupMeta;
-  final WidgetFactory wf;
 
-  BuildOp op;
+  late BuildOp op;
 
-  _TagTableGroup(this.wf, this.groupMeta, this.rows) {
+  _TagTableRowGroup(this.parent, this.groupMeta, this.rows) {
     op = BuildOp(onChild: onChild);
   }
 
@@ -209,57 +353,21 @@ class _TagTableGroup {
 
     final row = _TagTableDataRow();
     rows.add(row);
-    childMeta.register(_TagTableRow(wf, childMeta, row).op);
-  }
-}
-
-class _TagTableRow {
-  final _TagTableDataRow row;
-  final BuildMetadata rowMeta;
-  final WidgetFactory wf;
-
-  BuildOp op;
-  BuildOp _cellOp;
-
-  _TagTableRow(this.wf, this.rowMeta, this.row) {
-    op = BuildOp(onChild: onChild);
-  }
-
-  void onChild(BuildMetadata childMeta) {
-    if (childMeta.element.parent != rowMeta.element) return;
-    if (TagTable._getCssDisplayValue(childMeta) != kCssDisplayTableCell) {
-      return;
-    }
-
-    _cellOp ??= BuildOp(
-      onWidgets: (cellMeta, widgets) {
-        final column = wf
-            .buildColumnPlaceholder(cellMeta, widgets)
-            ?.wrapWith((_, child) => _TableCell(child));
-        if (column == null) return [];
-
-        final attributes = cellMeta.element.attributes;
-        row.cells.add(_TagTableDataCell(
-          child: column,
-          colspan: tryParseIntFromMap(attributes, 'colspan') ?? 1,
-          rowspan: tryParseIntFromMap(attributes, 'rowspan') ?? 1,
-        ));
-
-        return [column];
-      },
-      priority: StyleSizing.kPriority,
-    );
-
-    childMeta.register(_cellOp);
+    childMeta.register(_TagTableRow(parent, childMeta, row).op);
   }
 }
 
 @immutable
 class _TagTableData {
-  final captions = <Widget>[];
+  final bodies = <_TagTableDataGroup>[];
   final footer = _TagTableDataGroup();
   final header = _TagTableDataGroup();
-  final rows = <_TagTableDataRow>[];
+
+  _TagTableDataGroup get body {
+    final body = _TagTableDataGroup();
+    bodies.add(body);
+    return body;
+  }
 }
 
 @immutable
@@ -276,7 +384,9 @@ class _TagTableDataRow {
 class _TagTableDataCell {
   final Widget child;
   final int colspan;
+  final BuildMetadata meta;
   final int rowspan;
 
-  _TagTableDataCell({this.child, this.colspan, this.rowspan});
+  _TagTableDataCell(this.meta,
+      {required this.child, required this.colspan, required this.rowspan});
 }
