@@ -109,11 +109,6 @@ class TextStyleOps {
     );
   }
 
-  static BuildOp textDecoOp(TextDeco v) => BuildOp(
-        onTree: (meta, _) =>
-            meta.willBuildSubtree ? null : meta.tsb<TextDeco>(textDeco, v),
-      );
-
   static TextStyleHtml textDirection(TextStyleHtml p, String v) {
     final textDirection = (v == kCssDirectionRtl)
         ? TextDirection.rtl
@@ -135,14 +130,14 @@ class TextStyleOps {
     for (final part in parts) {
       final fontFamily = part
           .trim()
-          .replaceFirstMapped(RegExp(r"""^("|')(.+)\1$"""), (m) => m.group(2));
+          .replaceFirstMapped(RegExp(r"""^("|')(.+)\1$"""), (m) => m[2]!);
       if (fontFamily.isNotEmpty) list.add(fontFamily);
     }
 
     return list;
   }
 
-  static FontStyle fontStyleTryParse(String value) {
+  static FontStyle? fontStyleTryParse(String value) {
     switch (value) {
       case kCssFontStyleItalic:
         return FontStyle.italic;
@@ -153,7 +148,7 @@ class TextStyleOps {
     return null;
   }
 
-  static FontWeight fontWeightTryParse(String value) {
+  static FontWeight? fontWeightTryParse(String value) {
     switch (value) {
       case kCssFontWeightBold:
         return FontWeight.bold;
@@ -180,12 +175,13 @@ class TextStyleOps {
     return null;
   }
 
-  static double _fontSizeTryParse(WidgetFactory wf, TextStyleHtml p, String v) {
+  static double? _fontSizeTryParse(
+      WidgetFactory wf, TextStyleHtml p, String v) {
     final length = tryParseCssLength(v);
     if (length != null) {
       final lengthValue = length.getValue(
         p,
-        baseValue: p.parent?.style?.fontSize,
+        baseValue: p.parent?.style.fontSize,
         scaleFactor: p.getDependency<MediaQueryData>().textScaleFactor,
       );
       if (lengthValue != null) return lengthValue;
@@ -208,55 +204,58 @@ class TextStyleOps {
         return _fontSizeMultiplyRootWith(p, .5625);
 
       case kCssFontSizeLarger:
-        return _fontSizeMultiplyWith(p.parent?.style?.fontSize, 1.2);
+        return _fontSizeMultiplyWith(p.parent?.style.fontSize, 1.2);
       case kCssFontSizeSmaller:
-        return _fontSizeMultiplyWith(p.parent?.style?.fontSize, 15 / 18);
+        return _fontSizeMultiplyWith(p.parent?.style.fontSize, 15 / 18);
     }
 
     return null;
   }
 
-  static double _fontSizeMultiplyRootWith(TextStyleHtml tsh, double value) {
+  static double? _fontSizeMultiplyRootWith(TextStyleHtml tsh, double value) {
     var root = tsh;
     while (root.parent != null) {
-      root = root.parent;
+      root = root.parent!;
     }
 
     return _fontSizeMultiplyWith(root.style.fontSize, value);
   }
 
-  static double _fontSizeMultiplyWith(double fontSize, double value) =>
+  static double? _fontSizeMultiplyWith(double? fontSize, double value) =>
       fontSize != null ? fontSize * value : null;
 
-  static double _lineHeightTryParse(
+  static double? _lineHeightTryParse(
       WidgetFactory wf, TextStyleHtml p, String v) {
     if (v == kCssLineHeightNormal) return -1;
 
     final number = double.tryParse(v);
     if (number != null && number > 0) return number;
 
+    final fontSize = p.style.fontSize;
+    if (fontSize == null) return null;
+
     final length = tryParseCssLength(v);
     if (length == null) return null;
 
     final lengthValue = length.getValue(
       p,
-      baseValue: p.style.fontSize,
+      baseValue: fontSize,
       scaleFactor: p.getDependency<MediaQueryData>().textScaleFactor,
     );
     if (lengthValue == null) return null;
 
-    return lengthValue / p.style.fontSize;
+    return lengthValue / fontSize;
   }
 }
 
 @immutable
 class TextDeco {
-  final Color color;
-  final bool over;
-  final bool strike;
-  final TextDecorationStyle style;
-  final CssLength thickness;
-  final bool under;
+  final Color? color;
+  final bool? over;
+  final bool? strike;
+  final TextDecorationStyle? style;
+  final CssLength? thickness;
+  final bool? under;
 
   TextDeco({
     this.color,
@@ -267,9 +266,9 @@ class TextDeco {
     this.under,
   });
 
-  factory TextDeco.tryParse(String value) {
-    for (final v in splitCssValues(value)) {
-      switch (v) {
+  static TextDeco? tryParse(List<String> values) {
+    for (final value in values) {
+      switch (value) {
         case kCssTextDecorationLineThrough:
           return TextDeco(strike: true);
         case kCssTextDecorationNone:
