@@ -64,6 +64,7 @@ class _PostScreen extends StatelessWidget {
           ],
         ),
         body: Padding(
+          padding: const EdgeInsets.all(8),
           child: SingleChildScrollView(
             child: Column(
               children: [
@@ -79,7 +80,6 @@ class _PostScreen extends StatelessWidget {
               ],
             ),
           ),
-          padding: const EdgeInsets.all(8),
         ),
       );
 
@@ -104,8 +104,9 @@ class _PostsState extends State<_PostsList> {
   void initState() {
     super.initState();
 
-    posts =
-        http.get(widget.url).then((resp) => _parseJson(jsonDecode(resp.body)));
+    posts = http
+        .get(Uri.parse(widget.url))
+        .then((resp) => _parseJson(jsonDecode(resp.body)));
   }
 
   @override
