@@ -4,23 +4,25 @@ const kCssMargin = 'margin';
 
 Widget _marginHorizontalBuilder(Widget w, CssLengthBox b, TextStyleHtml tsh) =>
     Padding(
-      child: w,
       padding: EdgeInsets.only(
         left: b.getValueLeft(tsh) ?? 0.0,
         right: b.getValueRight(tsh) ?? 0.0,
       ),
+      child: w,
     );
 
 class StyleMargin {
+  static const kPriorityBoxModel9k = 9000;
+
   final WidgetFactory wf;
 
   StyleMargin(this.wf);
 
   BuildOp get buildOp => BuildOp(
         onTree: (meta, tree) {
-          if (meta.willBuildSubtree) return;
+          if (meta.willBuildSubtree == true) return;
           final m = tryParseCssLengthBox(meta, kCssMargin);
-          if (m?.hasLeftOrRight != true) return;
+          if (m == null || !m.hasLeftOrRight) return;
 
           return wrapTree(
             tree,
@@ -29,33 +31,25 @@ class StyleMargin {
           );
         },
         onWidgets: (meta, widgets) {
-          if (widgets?.isNotEmpty != true) return null;
+          if (widgets.isEmpty) return null;
           final m = tryParseCssLengthBox(meta, kCssMargin);
           if (m == null) return null;
 
-          final t = m.top?.isNotEmpty == true;
-          final b = m.bottom?.isNotEmpty == true;
-          final ws = List<WidgetPlaceholder>.filled(
-              (t ? 1 : 0) + widgets.length + (b ? 1 : 0), null);
-          final tsb = meta.tsb();
+          final tsb = meta.tsb;
 
-          var i = 0;
-          if (t) ws[i++] = HeightPlaceholder(m.top, tsb);
-
-          for (final widget in widgets) {
-            if (m.hasLeftOrRight) {
-              widget.wrapWith(
-                  (c, w) => _marginHorizontalBuilder(w, m, tsb.build(c)));
-            }
-
-            ws[i++] = widget;
-          }
-
-          if (b) ws[i++] = HeightPlaceholder(m.bottom, tsb);
-
-          return ws;
+          return [
+            if (m.top?.isNotEmpty ?? false) HeightPlaceholder(m.top!, tsb),
+            for (final widget in widgets)
+              if (m.hasLeftOrRight)
+                widget.wrapWith(
+                    (c, w) => _marginHorizontalBuilder(w, m, tsb.build(c)))
+              else
+                widget,
+            if (m.bottom?.isNotEmpty ?? false)
+              HeightPlaceholder(m.bottom!, tsb),
+          ];
         },
         onWidgetsIsOptional: true,
-        priority: 99999,
+        priority: kPriorityBoxModel9k,
       );
 }
