@@ -10,19 +10,21 @@ WidgetPlaceholder _paddingInlineBefore(TextStyleBuilder tsb, CssLengthBox b) =>
     WidgetPlaceholder<CssLengthBox>(b).wrapWith((context, _) =>
         _paddingInlineSizedBox(b.getValueLeft(tsb.build(context))));
 
-Widget _paddingInlineSizedBox(double width) =>
+Widget _paddingInlineSizedBox(double? width) =>
     width != null && width > 0 ? SizedBox(width: width) : widget0;
 
 class StylePadding {
+  static const kPriorityBoxModel3k = 3000;
+
   final WidgetFactory wf;
 
   StylePadding(this.wf);
 
   BuildOp get buildOp => BuildOp(
         onTree: (meta, tree) {
-          if (meta.willBuildSubtree) return;
+          if (meta.willBuildSubtree == true) return;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
-          if (padding?.hasLeftOrRight != true) return;
+          if (padding == null || !padding.hasLeftOrRight) return;
 
           return wrapTree(
             tree,
@@ -33,7 +35,7 @@ class StylePadding {
           );
         },
         onWidgets: (meta, widgets) {
-          if (widgets?.isNotEmpty != true) return null;
+          if (widgets.isEmpty) return null;
           final padding = tryParseCssLengthBox(meta, kCssPadding);
           if (padding == null) return null;
 
@@ -42,12 +44,12 @@ class StylePadding {
               ?.wrapWith((c, w) => _build(c, meta, w, padding)));
         },
         onWidgetsIsOptional: true,
-        priority: 9999,
+        priority: kPriorityBoxModel3k,
       );
 
-  Widget _build(BuildContext context, BuildMetadata meta, Widget child,
+  Widget? _build(BuildContext context, BuildMetadata meta, Widget child,
       CssLengthBox padding) {
-    final tsh = meta.tsb().build(context);
+    final tsh = meta.tsb.build(context);
     return wf.buildPadding(
       meta,
       child,
